More structured fdt reader

+++ src/arch/riscv/boot.rs
@@ -3,7 +3,7 @@ use core::arch::global_asm;
 global_asm!(include_str!("entry.s"));
 
 #[unsafe(no_mangle)]
-extern "C" fn rust_entry(hart_id: u64, fdt: *const u32) -> ! {
+extern "C" fn rust_entry(hart_id: u64, fdt: *const u8) -> ! {
     crate::arch::println!("We are running on hart {hart_id}");
     super::device_tree::parse(fdt);
     crate::main();
